refactor: count 쿼리 반환 타입 변경 및 Link 테이블 인덱스 추가 (#207)

+++ src/main/java/com/sofa/linkiving/domain/link/dto/response/LinkTotalCountRes.java
@@ -4,6 +4,6 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 public record LinkTotalCountRes(
 	@Schema(description = "링크 전체 개수")
-	int totalCount
+	Long totalCount
 ) {
 }
